Moved js execution methods to engine dependency
