+++ kotest-framework/kotest-framework-api/build.gradle.kts
@@ -41,8 +41,6 @@ kotlin {
       val jsMain by getting {
          dependsOn(commonMain)
          dependencies {
-            // this must be api as it's compiled into the final source
-            api(kotlin("test-js"))
             implementation(Libs.Coroutines.coreJs)
          }
       }
