+++ kotest-framework/kotest-framework-engine/src/jsMain/kotlin/io/kotest/engine/execution/jasmineTestMethods.kt
@@ -1,8 +1,10 @@
-package io.kotest.core.js
+package io.kotest.engine.execution
 
 /**
  * Defines the jasmine/mocha/karma style test function names as external functions.
  * Then we can invoke them from the test engine.
+ *
+ * Note: At runtime, one of the supported JS test frameworks must make these functions available.
  */
 external fun describe(name: String, fn: () -> Unit)
 external fun xdescribe(name: String, fn: () -> Unit)
